100% test coverage 🎉
